# docs/api.yml
openapi: "3.0.2"
info:
  title: Scholarship API
  version: "1.0"
servers:
  - url: http://localhost:8080
  - url: https://virtserver.swaggerhub.com/AmourRamanantsiresy/Scholarship/1.0.0
paths:
...
  /university:
    get:
      tags:
        - "University"
      operationId: getAllUniversities
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
        - $ref: '#/components/parameters/PUniversity'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestUniversity'
    put:
      tags:
        - University
      operationId: createOrUpdateUniversity
      description: Update an existing university if you specify the id else create new university
      requestBody:
        content:
          application/json:
            schema: 
              type: array
              items:
                $ref: '#/components/schemas/CreateUniversity'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestUniversity'
  /university/search:
    get:
      tags:
        - "University"
      operationId: searchUniversities
      parameters:
        - $ref: '#/components/parameters/PSearch'
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
        - $ref: '#/components/parameters/PUniversity'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestUniversity'
...
components:
  schemas:
...
    IdName:
      type: object
      properties:
        id: 
          type: integer
        name: 
          type: string
...
    CreateUniversity:
      properties:
        country:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
      allOf:
        - $ref: '#/components/schemas/IdName'
...
    RestUniversity:
      properties:
        country:
          type: string
          example: America
        languages: 
          type: array
          items:
            type: string
            example: english
      allOf:
        - $ref: '#/components/schemas/IdName'
...
  parameters:
    PSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
    PPage:
      name: page
      in: query
      required: false
      schema:
        type: integer
    POrder:
      name: order
      in: query
      required: false
      schema:
        enum: [ASC, DESC]
    PSearch:
      name: value
      in: query
      required: true
      schema:
        type: string
...
    PUniversity:
      name: by
      in: query
      required: false
      schema:
        type: string
        enum: [LANGUAGE, COUNTRY, NAME]
...